requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-jazzmin==2.6.0
django-shortcuts==1.6
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
